compile benchs and test

+++ pallets/parachain-staking/src/tests.rs
@@ -6720,7 +6720,7 @@ fn deferred_payment_steady_state_event_flow() {
 						total_exposed_amount: 400,
 					},
 					Event::NewRound {
-						starting_block: (round as u64 - 1) * 5,
+						starting_block: (round as u32 - 1) * 5,
 						round: round as u32,
 						selected_collators_number: 4,
 						total_balance: 1600,
